pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "friday-gemini"
version = "0.1.0"
description = "Assistente pessoal para Windows com Gemini e automacao de apps"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.32.3",
  "python-dotenv>=1.0.1",
  "pyautogui>=0.9.54",
  "SpeechRecognition>=3.10.4",
  "pyttsx3>=2.98",
  "PyAudioWPatch>=0.2.12.7",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
